第二次优化，主要是图片操作的封装和代码重构
