+++ Demo/src/main/java/com/VanGogh/demo/Entities/UserLoginStatusEntity.java
@@ -9,15 +9,13 @@ import java.time.LocalDateTime;
 
 @Entity
 @Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserLoginStatusEntity {
+public class UserLoginStatusEntity{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false)
+    @Column
     private LocalDateTime loginTime;
 
     @OneToOne
